update products table.

+++ db/migrate/20180430212355_add_products.rb
@@ -2,7 +2,7 @@ class AddProducts < ActiveRecord::Migration[5.2]
   def change
     create_table :products do |t|
       t.column :name, :varchar
-      t.column :price, :integer
+      t.column :price, :money
 
       t.timestamps
     end
